Airspace display is now kept also to lowest zoom distances where airport diagram shows up. Fixed warnings about integer conversion.

+++ src/mappainter/mappainter.h
@@ -305,7 +305,7 @@ protected:
   void drawText(Marble::GeoPainter *painter, const atools::geo::Pos& pos, const QString& text, bool topCorner, bool leftCorner);
 
   /* Drawing functions for simple geometry */
-  void drawCircle(Marble::GeoPainter *painter, const atools::geo::Pos& center, int radius);
+  void drawCircle(Marble::GeoPainter *painter, const atools::geo::Pos& center, float radius);
   void drawCross(Marble::GeoPainter *painter, int x, int y, int size);
 
   /* No GC and no rhumb */
